:globe_with_meridians: New RU locale lines from issue #1654 (#1658)
## Translations
- [x] `.json` files are formatted: `yarn format`
- [x] Translations are correct

This pull request adds new RU locale lines to the cad-settings.json
file, addressing issue #1654. The changes were made to improve the user
experience for Russian-speaking users by adding previously
not-localizable lines.

Please note that this is the first step in addressing the issue, and
future updates to the locale may be necessary. (the future lines will be
in the following PR after their implementation in the main branch)

---------

Co-authored-by: Dev-CasperTheGhost <53900565+Dev-CasperTheGhost@users.noreply.github.com>