🌐 chore: Russian translation update (#1670)
## Translations Chore for RU

- [x] `.json` files are formatted: `yarn format`
- [x] Translations are correct

This pull request adds new locale strings for the Russian language that
were introduced in the 1.59.0 update. The new strings have been added to
the appropriate language files and have been tested to ensure they
display correctly in the application.

Please review and merge this pull request if everything looks good.

Thanks!