🌐 chore: Updating RU translations (#1684)
Added translations for RU locale for callsigns update
- [x] `.json` files are formatted: `yarn format`
- [x] Translations are correct

This PR adds translations for the RU locale that updates the callsigns
strings in the files. The changes were made to improve the user
experience for Russian-speaking users.

Please review the changes and let me know if there are any issues or
concerns. Thank you for your time and attention to this repo.